vision sim heat
