+++ src/main/kotlin/com/team4099/robot2025/subsystems/vision/Vision.kt
@@ -8,19 +8,20 @@ import com.team4099.robot2025.config.constants.FieldConstants
 import com.team4099.robot2025.config.constants.VisionConstants
 import com.team4099.robot2025.subsystems.superstructure.Request
 import com.team4099.robot2025.subsystems.vision.camera.CameraIO
-import com.team4099.robot2025.subsystems.vision.camera.CameraIOPVSim
 import com.team4099.robot2025.util.FMSData
 import com.team4099.robot2025.util.toTransform3d
 import edu.wpi.first.apriltag.AprilTagFieldLayout
 import edu.wpi.first.apriltag.AprilTagFields
 import edu.wpi.first.math.VecBuilder
+import edu.wpi.first.math.geometry.Pose2d
 import edu.wpi.first.math.geometry.Rotation2d
 import edu.wpi.first.wpilibj.DriverStation
 import edu.wpi.first.wpilibj.RobotBase
 import edu.wpi.first.wpilibj2.command.SubsystemBase
+import java.util.function.Supplier
 import org.littletonrobotics.junction.Logger
 import org.photonvision.PhotonUtils
-import org.team4099.lib.geometry.Pose2d
+import org.photonvision.simulation.VisionSystemSim
 import org.team4099.lib.geometry.Pose3d
 import org.team4099.lib.geometry.Rotation3d
 import org.team4099.lib.geometry.Transform2d
@@ -37,13 +38,9 @@ import org.team4099.lib.units.derived.cos
 import org.team4099.lib.units.derived.degrees
 import org.team4099.lib.units.derived.inRadians
 import org.team4099.lib.units.derived.sin
-import java.util.function.Consumer
-import java.util.function.Supplier
-import org.photonvision.proto.Photon
-import org.photonvision.simulation.PhotonCameraSim
-import org.photonvision.simulation.VisionSystemSim
 
-class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.math.geometry.Pose2d>) : SubsystemBase() {
+class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<Pose2d> = Supplier { Pose2d() }) :
+  SubsystemBase() {
   val io: List<CameraIO> = cameras.toList()
   val inputs = List(io.size) { CameraIO.CameraInputs() }
 
@@ -64,8 +61,6 @@ class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.
   var isAutoAligning = false
   var isAligned = false
 
-  var drivetrainOdometry: () -> Pose2d = { Pose2d() }
-
   private val xyStdDev = TunableNumber("Vision/xystdev", VisionConstants.XY_STDDEV)
 
   private val thetaStdDev = TunableNumber("Vision/thetaStdDev", VisionConstants.THETA_STDDEV)
@@ -76,10 +71,6 @@ class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.
 
   var lastTrigVisionUpdate =
     TimestampedTrigVisionUpdate(Clock.fpgaTime, -1, Transform2d(Translation2d(), 0.degrees))
-
-  private var fieldFramePoseSupplier = Supplier<Pose2d> { Pose2d() }
-  private var visionConsumer: Consumer<List<TimestampedVisionUpdate>> = Consumer {}
-  private var reefVisionConsumer: Consumer<TimestampedTrigVisionUpdate> = Consumer {}
 
   private var lastSeenTagId: Int? = null
   private var pulseEndTime = 0.0.seconds
@@ -93,18 +84,10 @@ class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.
       visionSim = VisionSystemSim("main")
       visionSim!!.addAprilTags(AprilTagFieldLayout.loadField(AprilTagFields.kDefaultField))
 
-      cameras.forEach { camera -> visionSim!!.addCamera(camera.cameraSim, camera.transform.transform3d)}
+      cameras.forEach { camera ->
+        visionSim!!.addCamera(camera.cameraSim, camera.transform.transform3d)
+      }
     }
-  }
-
-  fun setDataInterfaces(
-    fieldFramePoseSupplier: Supplier<Pose2d>,
-    visionConsumer: Consumer<List<TimestampedVisionUpdate>>,
-    reefVisionMeasurementConsumer: Consumer<TimestampedTrigVisionUpdate>
-  ) {
-    this.fieldFramePoseSupplier = fieldFramePoseSupplier
-    this.visionConsumer = visionConsumer
-    this.reefVisionConsumer = reefVisionMeasurementConsumer
   }
 
   override fun periodic() {
@@ -159,14 +142,6 @@ class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.
             if ((tag.fiducialId in VisionConstants.BLUE_REEF_TAGS && FMSData.isBlue) ||
               (tag.fiducialId in VisionConstants.RED_REEF_TAGS && !FMSData.isBlue)
             ) {
-
-              val aprilTagAlignmentAngle =
-                if (FMSData.isBlue) {
-                  VisionConstants.BLUE_REEF_TAG_THETA_ALIGNMENTS[tag.fiducialId]
-                } else {
-                  VisionConstants.RED_REEF_TAG_THETA_ALIGNMENTS[tag.fiducialId]
-                }
-
               val fieldTTag =
                 FieldConstants.AprilTagLayoutType.OFFICIAL
                   .layout
@@ -325,16 +300,9 @@ class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.
                 closestReefTagAcrossCams?.value?.second?.rotation?.z ?: 0.degrees
               )
             )
-
-          // reefVisionConsumer.accept(lastTrigVisionUpdate)
         }
       }
     }
-
-    // visionConsumer.accept(visionUpdates)
-    Logger.recordOutput(
-      "LoggedRobot/VisionLoopTimeMS", (Clock.realTimestamp - startTime).inMilliseconds
-    )
 
     val now = Clock.fpgaTime
 
@@ -362,6 +330,10 @@ class Vision(vararg cameras: CameraIO, val poseSupplier: Supplier<edu.wpi.first.
     if (now > pulseEndTime) {
       autoAlignReadyRumble = false
     }
+
+    Logger.recordOutput(
+      "LoggedRobot/VisionLoopTimeMS", (Clock.realTimestamp - startTime).inMilliseconds
+    )
   }
 
   companion object {
